added names
I added the names and courses of the buttfucks. Clicking on the prof
will bring you to rateyourprof and clicking on the course name will
bring you uwflow course rating. In the future, clicking the exam type
will bring you to a site that will help you study that topic. I will
also have a voting system on the right side of each thing